Add useModal
