+++ src/components/hooks/useModal.js
@@ -1,0 +1,19 @@
+import { useState, useCallback } from 'react';
+
+export const useModal = () => {
+	const [isModalOpen, setIsModalOpen] = useState(false);
+
+	const openModal = useCallback(() => {
+		setIsModalOpen(true);
+	}, []);
+
+	const closeModal = useCallback(() => {
+		setIsModalOpen(false);
+	}, []);
+
+	return {
+		isModalOpen,
+		openModal,
+		closeModal,
+	};
+};
